requirements.txt
streamlit>=1.33
llama-index>=0.10
openai>=1.3
pyyaml
tiktoken
